feat: multiple features, fixes, and refactors from supermarket sweep (#24)
* chore: better comment on max file name len

* feat: get meshbin struct size reduced, cleanup load queue as its processed

* refactor(loading): into a switch/case

* feat(gameobjects): dump gameobjects when loading with dumpExisting

* fix/refactor: free textures from ram, reduce max bones to 15, increase max loaded meshes to 100

* refactor: ditch krom font, give more space to bump allocator

* refactor: increase ot size to 2k, no need to do loading queue backwards

* feat: add simple fade-to-black fog, update tim tool to specify transparency colour

* refactor(fog): add a min distance before fog starts to kick in

* refactor(fog): apply fade-to-black fog to billboards and sprites

* feat(render): add the ability to override what gameobjects are rendered via

* refactor(perf): more accurate fps counter in perf monitor

* feat: make sure all corners+centre of object are within range before looking at faces, fix ot size

* refactor(renderer): add render flags to game objects

* feat(render): use sphereboundingradius to determine if a gameobject is on screen before doing face rendering

* refactor(render): rename some constexpr, read sz0-4 straight into function

* refactor(renderer): dont use cam delta in objectvisible check

* feat(rand): add more rand options

* refactor(renderer): minor changes

* fix(menu): crash when handling inputs on empty menu, can now set renderer clear colour

* feat: add lerp functions, improve sprite hud element, tim creator

* fix(renderer): fog distance check applying when fog wasnt enabled

* feat(spu): first iteration of loading vag files

* fix(spu): loading/playing bugs in the sound manager

* chore(spu): add max channel id, clamp to that when playing

* refactor(sound): add a channelplaybackconfig helper

* refactor(mesh): only assign skeleton data when needed, not all the time

* refactor(renderer): handle updated mesh struct

* chore: add missed bits from renderer update

* feat(mesh): add bounding sphere back in

* feat(renderer): check if game object is visible before rendering

* fix(mesh): bounding sphere generation

* feat(tools): generate bounding sphere radius when generating meshbin

* refactor: much improved bounding sphere radius logic, generated by the tools

* feat(render): subdivision of textured quads

* feat(render): support tris at last! (including sub division)

* refactor(render): dont suvdivide prim if its leaving screen space because it looks worse

+++ src/ui/hud/sprite_hud_element.cpp
@@ -18,6 +18,6 @@ void SpriteHUDElement::Render(const psyqo::Rect &parentRect)
     if (!m_isEnabled)
         return;
 
-    psyqo::Rect rect = {.pos = {parentRect.pos.x + m_rect.pos.x, parentRect.pos.y + m_rect.pos.y}, .size = m_rect.size};
+    psyqo::Rect rect = {.pos = {static_cast<int16_t>(parentRect.pos.x + m_rect.pos.x), static_cast<int16_t>(parentRect.pos.y + m_rect.pos.y)}, .size = m_rect.size};
     Renderer::Instance().RenderSprite(m_tim, rect, m_spriteUV);
 }
